+++ build.gradle
@@ -11,6 +11,7 @@ buildscript {
 
 plugins {
     id 'java'
+    id 'maven'
     id 'maven-publish'
     id 'checkstyle'
     id "net.researchgate.release" version "2.6.0"
