--- src/main/java/org/motometer/telegram/bot/TelegramApiException.java
@@ -1,12 +1,0 @@
-package org.motometer.telegram.bot;
-
-public class TelegramApiException extends RuntimeException {
-
-    public TelegramApiException(Throwable e) {
-        super(e);
-    }
-
-    public TelegramApiException(String message) {
-        super(message);
-    }
-}
