--- src/main/java/org/motometer/telegram/bot/UpdateListener.java
@@ -1,8 +1,0 @@
-package org.motometer.telegram.bot;
-
-import org.motometer.telegram.bot.api.Update;
-
-public interface UpdateListener {
-
-    void onUpdate(Update update);
-}
