+++ src/main/java/org/motometer/telegram/bot/api/Bot.java
@@ -1,8 +1,8 @@
-package org.motometer.telegram.bot;
+package org.motometer.telegram.bot.api;
 
 import org.motometer.telegram.bot.api.message.Message;
-import org.motometer.telegram.bot.api.Update;
-import org.motometer.telegram.bot.api.User;
+import org.motometer.telegram.bot.api.update.Update;
+import org.motometer.telegram.bot.api.update.User;
 import org.motometer.telegram.bot.api.method.SendMessage;
 
 import java.util.List;
@@ -19,18 +19,18 @@ public interface Bot {
      * @return {@link User}
      * @see <a href="https://core.telegram.org/bots/api#getme">getMe</a>
      */
-    User me();
+    User me() throws BotException;
 
     /**
      * @return {@link Update}
      * @see <a href="https://core.telegram.org/bots/api#getupdates">getUpdates</a>
      */
-    List<Update> updates();
+    List<Update> updates() throws BotException;
 
     /**
      * @param message request entity
      * @return {@link Message}
      * @see <a href="https://core.telegram.org/bots/api#sendmessage">sendMessage</a>
      */
-    Message sendMessage(SendMessage message);
+    Message sendMessage(SendMessage message) throws BotException;
 }
