+++ src/main/java/org/motometer/telegram/bot/api/package-info.java
@@ -1,4 +1,1 @@
-@Gson.TypeAdapters
 package org.motometer.telegram.bot.api;
-
-import org.immutables.gson.Gson;
