+++ src/main/java/org/motometer/telegram/bot/api/update/ChatType.java
@@ -1,4 +1,4 @@
-package org.motometer.telegram.bot.api;
+package org.motometer.telegram.bot.api.update;
 
 import lombok.RequiredArgsConstructor;
