+++ src/main/java/org/motometer/telegram/bot/api/update/package-info.java
@@ -1,0 +1,4 @@
+@Gson.TypeAdapters
+package org.motometer.telegram.bot.api.update;
+
+import org.immutables.gson.Gson;
