--- src/main/java/org/motometer/telegram/bot/package-info.java
@@ -1,1 +1,0 @@
-package org.motometer.telegram.bot;
