--- src/test/java/org/motometer/telegram/api/package-info.java
@@ -1,1 +1,0 @@
-package org.motometer.telegram.api;
