+++ src/test/java/org/motometer/telegram/bot/api/update/package-info.java
@@ -1,0 +1,1 @@
+package org.motometer.telegram.bot.api.update;
